+++ src/dict_from_g2pE/core.py
@@ -1,16 +1,17 @@
 from collections import OrderedDict
 from functools import partial
+from logging import getLogger
 from multiprocessing.pool import Pool
-from typing import Dict, Optional, Set, Tuple
+from typing import Any, Dict, Optional, Set, Tuple
 
-from g2p_en import G2p
+import nltk
 from ordered_set import OrderedSet
 from pronunciation_dictionary import PronunciationDict, Pronunciations, Word
 from tqdm import tqdm
 from word_to_pronunciation import Options, get_pronunciations_from_word
 
 
-def transcribe_with_g2pE(vocabulary: OrderedSet[Word], weight: float = 1.0, trim: Optional[Set[str]] = None, split_on_hyphen: bool = True, n_jobs: int = 4, maxtasksperchild: Optional[int] = None, chunksize: int = 100_000, silent: bool = False):
+def transcribe_with_g2pE(vocabulary: OrderedSet[Word], *, weight: float = 1.0, trim: Optional[Set[str]] = None, split_on_hyphen: bool = True, n_jobs: int = 4, maxtasksperchild: Optional[int] = None, chunksize: int = 100_000, silent: bool = False):
   if trim is None:
     trim = set()
   trim_symbols = ''.join(trim)
@@ -20,6 +21,20 @@ def transcribe_with_g2pE(vocabulary: OrderedSet[Word], weight: float = 1.0, trim
   return dictionary_instance
 
 
+def download_nltk_data():
+  logger = getLogger(__name__)
+  try:
+    nltk.data.find('taggers/averaged_perceptron_tagger.zip')
+  except LookupError:
+    logger.info("Downloading 'averaged_perceptron_tagger' from nltk ...")
+    nltk.download('averaged_perceptron_tagger')
+  try:
+    nltk.data.find('corpora/cmudict.zip')
+  except LookupError:
+    logger.info("Downloading 'cmudict' from nltk ...")
+    nltk.download('cmudict')
+
+
 def get_pronunciations(vocabulary: OrderedSet[Word], weight: float, options: Options, n_jobs: int, maxtasksperchild: Optional[int], chunksize: int, silent: bool) -> PronunciationDict:
   lookup_method = partial(
     process_get_pronunciation,
@@ -27,6 +42,10 @@ def get_pronunciations(vocabulary: OrderedSet[Word], weight: float, options: Opt
     options=options,
   )
 
+  download_nltk_data()
+  # G2p downloads NLTK data on import which could take a while
+  # pylint: disable=import-outside-toplevel
+  from g2p_en import G2p
   model = G2p()
 
   with Pool(
@@ -55,10 +74,10 @@ def get_dictionary(pronunciations_to_i: Dict[int, Pronunciations], vocabulary: O
 
 
 process_unique_words: OrderedSet[Word] = None
-process_model: G2p = None
+process_model: Any = None
 
 
-def __init_pool_prepare_cache_mp(words: OrderedSet[Word], model: G2p) -> None:
+def __init_pool_prepare_cache_mp(words: OrderedSet[Word], model: Any) -> None:
   global process_unique_words
   global process_model
   process_unique_words = words
@@ -84,7 +103,7 @@ def process_get_pronunciation(word_i: int, weight: float, options: Options) -> T
   return word_i, pronunciations
 
 
-def lookup_in_model(word: Word, model: G2p, weight: float) -> Pronunciations:
+def lookup_in_model(word: Word, model: Any, weight: float) -> Pronunciations:
   assert len(word) > 0
   # lower() because G2p seems to predict only lower-case words correctly
   word = word.lower()