+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- timepoint for downloading nltk data
+
 ## [0.0.2] - 2024-01-24
 
 ### Added